+++ DependencyInjection/LinkinSwaggerResolverExtension.php
@@ -37,7 +37,7 @@ class LinkinSwaggerResolverExtension extends Extension
         $config = $this->processConfiguration($configuration, $configs);
 
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('services.yml');
+        $loader->load('services.yaml');
 
         $this->registerConfigurationLoader($container, $config);
     }