+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [Unreleased]
+### Added
+- Added possibility for creating `SwaggerResolver` object for all defined swagger request parameters.
+### Changed
+- Moved `SwaggerResolver` building process into separate service `SwaggerResolverBuilder`.
+- Renamed `services.yml` into `services.yaml`.
+
 ## [0.1.2] - 2018-10-17
 ### Added
 - Added correct processing of the objects references.